double type change

+++ src/Basic_Calculator.java
@@ -24,7 +24,7 @@ public static void main(String [] args){
     System.out.println("Enter Your First Number:");
 
     //Input for first Integer
-    int a = user_input.nextInt();
+    double a = user_input.nextDouble();
 
     //Input for Operation
     System.out.println("Enter Operators (Eg : Sum(+) , Diff(-) , Multiplication( *) , Divison (/ ), Modulus (%) , Power(^)");
@@ -32,7 +32,7 @@ public static void main(String [] args){
 
     //Input for second Integer
     System.out.println("Enter Your Second Number:");
-    int b = user_input.nextInt();
+    double b = user_input.nextDouble();
     user_input.close();
 
     //Print new line to 
@@ -43,27 +43,27 @@ public static void main(String [] args){
     switch(o){
         
         case '+':
-        int sum = a+b;
+        double sum = a+b;
         System.out.println( "ANSWER:" +"     "+ a +"     "+ "+" + "       "+ b +"       "+"=" +"       "+ sum);
         break;
 
         case '-':
-        int sub = a-b;
+        double sub = a-b;
         System.out.println("ANSWER:" +"       " + a +"       " + "-" +"       " + b +"       " +"=" +"       " + sub);
         break;
 
         case '*':
-        int mul = a*b;
+        double mul = a*b;
         System.out.println("ANSWER:" +"       " +a +"       " + "*" +"       " +  b +"       "+"=" +"       " + mul);
         break;
 
         case '/':
-        int div = a/b;
+        double div = a/b;
         System.out.println("ANSWER:"+"       " +a +"       " + "/" +"       " + b +"       "+"=" +"       "+ div);
         break;
 
         case '%':
-        int mod = a%b;
+        double mod = a%b;
         System.out.println("ANSWER:"+"       " +a +"       " + "%" +"       " + b +"       "+"=" +"       " + mod);
 
         case '^':
